# api/openapi/openapi.json
{
  "components": {
...
    "responses": {
...
      "DansError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "DANS rejected or could not complete the request."
      },
...
    },
    "schemas": {
...
      "Comment": {
        "additionalProperties": false,
        "description": "A comment about an RRSet.",
        "properties": {
          "account": {
            "description": "Name of an account that added the comment",
            "type": "string"
          },
          "content": {
            "description": "The actual comment",
            "type": "string"
          },
          "modified_at": {
            "description": "Timestamp of the last change to the comment",
            "type": "integer"
          }
        },
        "title": "Comment",
        "type": "object"
      },
...
      "Error": {
        "additionalProperties": false,
        "description": "Returned when the server encounters an error, either in client input or internally",
        "properties": {
          "error": {
            "description": "A human readable error message",
            "type": "string"
          },
          "errors": {
            "description": "Optional array of multiple errors encountered during processing",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "error"
        ],
        "title": "Error",
        "type": "object"
      },
...
      "NullableString": {
        "type": [
          "string",
          "null"
        ]
      },
      "NullableTimestamp": {
        "format": "date-time",
        "type": [
          "string",
          "null"
        ]
      },
      "RRSet": {
        "additionalProperties": false,
        "description": "This represents a Resource Record Set (all records with the same name and type).",
        "properties": {
          "changetype": {
            "description": "MUST be added when updating the RRSet. Must be REPLACE or DELETE. With DELETE, all existing RRs matching name and type will be deleted, including all comments. With REPLACE: when records is present, all existing RRs matching name and type will be deleted, and then new records given in records will be created. If no records are left, any existing comments will be deleted as well. When comments is present, all existing comments for the RRs matching name and type will be deleted, and then new comments given in comments will be created.",
            "type": "string"
          },
          "comments": {
            "description": "List of Comment. Must be empty when changetype is set to DELETE. An empty list results in deletion of all comments. modified_at is optional and defaults to the current server time.",
            "items": {
              "$ref": "#/components/schemas/Comment"
            },
            "type": "array"
          },
          "name": {
            "description": "Name for record set (e.g. “www.powerdns.com.”)",
            "type": "string"
          },
          "records": {
            "description": "All records in this RRSet. When updating Records, this is the list of new records (replacing the old ones). Must be empty when changetype is set to DELETE. An empty list results in deletion of all records (and comments).",
            "items": {
              "$ref": "#/components/schemas/Record"
            },
            "type": "array"
          },
          "ttl": {
            "description": "DNS TTL of the records, in seconds. MUST NOT be included when changetype is set to “DELETE”.",
            "type": "integer"
          },
          "type": {
            "description": "Type of this record (e.g. “A”, “PTR”, “MX”)",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "ttl",
          "records"
        ],
        "title": "RRSet",
        "type": "object"
      },
...
      "RRsetBrowsePage": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/NullableString"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RRSet"
            },
            "maxItems": 100,
            "type": "array"
          },
          "last_refreshed_at": {
            "$ref": "#/components/schemas/NullableTimestamp"
          },
          "next_cursor": {
            "$ref": "#/components/schemas/NullableString"
          },
          "state": {
            "enum": [
              "indexing",
              "ready",
              "refreshing",
              "stale"
            ],
            "type": "string"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "state",
          "last_refreshed_at",
          "error"
        ],
        "type": "object"
      },
      "Record": {
        "additionalProperties": false,
        "description": "The RREntry object represents a single record.",
        "properties": {
          "content": {
            "description": "The content of this record",
            "type": "string"
          },
          "disabled": {
            "description": "Whether or not this record is disabled. When unset, the record is not disabled",
            "type": "boolean"
          },
          "modified_at": {
            "description": "Timestamp of the last record change.",
            "type": "integer"
          }
        },
        "required": [
          "content"
        ],
        "title": "Record",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "A DANS opaque API token. The PowerDNS upstream key is never accepted from clients.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BrowserSession": {
        "description": "Opaque browser session, current token and identity checks apply.",
        "in": "cookie",
        "name": "__Host-dans_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The DANS management API and its pinned PowerDNS Authoritative 5.1.3 compatibility surface.",
    "license": {
      "name": "MIT"
    },
    "title": "DANS and PowerDNS Authoritative HTTP API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/dans/servers/{server_id}/zones/{zone_id}/rrsets": {
      "get": {
        "operationId": "browseRRsets",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "maxLength": 1024,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "maxLength": 1024,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "match",
            "schema": {
              "enum": [
                "exact",
                "prefix"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "maxLength": 32,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "maxLength": 4096,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RRsetBrowsePage"
                }
              }
            },
            "description": "Complete RRsets from a published browsing generation."
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RRsetBrowsePage"
                }
              }
            },
            "description": "Initial indexing is in progress."
          },
          "default": {
            "$ref": "#/components/responses/DansError"
          }
        },
        "x-dans-access-class": "authenticated-read"
      }
    },
    "/dans/servers/{server_id}/zones/{zone_id}/rrsets/refresh": {
      "post": {
        "operationId": "refreshRRsets",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "maxLength": 1024,
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RRsetBrowsePage"
                }
              }
            },
            "description": "Read refresh requested; no DNS mutation is performed."
          },
          "default": {
            "$ref": "#/components/responses/DansError"
          }
        },
        "x-dans-access-class": "authenticated-read"
      }
    },
...
  },
  "security": [
    {
      "APIKeyHeader": []
    },
    {
      "APIKeyHeader": []
    },
    {
      "BrowserSession": []
    }
  ],
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}